VW-932 Create OrderMaterial list

+++ Infrastructure/ServiceCore/Model/PageData.cs
@@ -24,10 +24,15 @@ namespace VErp.Infrastructure.ServiceCore.Model
 
     public class PageDataTable
     {
-        public int Total { get; set; }
+        public long Total { get; set; }
         public IList<NonCamelCaseDictionary> List { get; set; }
 
-        public static implicit operator PageDataTable((DataTable list, int total) result)
+        //public static implicit operator PageDataTable((DataTable list, int total) result)
+        //{
+        //    return (result.list, result.total);
+        //}
+
+        public static implicit operator PageDataTable((DataTable list, long total) result)
         {
             var lst = new List<NonCamelCaseDictionary>();
